bybit-importer/requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.6
aiohttp==3.10.11
aiosignal==1.3.2
attrs==25.1.0
ccxt==4.4.59
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudpickle==3.1.1
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
dnspython==2.7.0
Farama-Notifications==0.0.4
filelock==3.17.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
gymnasium==1.0.0
idna==3.10
importlib==1.0.4
Jinja2==3.1.5
kiwisolver==1.4.8
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mongoengine==0.29.1
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==2.2.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pandas==2.2.3
pillow==11.1.0
ppo==0.1.1
propcache==0.2.1
pycares==4.5.0
pycparser==2.22
pymongo==4.11
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
setuptools==75.8.0
six==1.17.0
stable_baselines3==2.5.0
sympy==1.13.1
torch==2.6.0
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
yarl==1.18.3
